+++ src/org/globantUniversity/data/teacher/TeacherPartTime.java
@@ -2,14 +2,14 @@ package org.globantUniversity.data.teacher;
 
 public class TeacherPartTime extends Teacher{
     private double baseSalary;
-    private byte activeHoursPerWeek;
-    public TeacherPartTime(String name, byte activeHoursPerWeek){
+    private int activeHoursPerWeek;
+    public TeacherPartTime(String name, int activeHoursPerWeek){
         super(name);
         this.activeHoursPerWeek = activeHoursPerWeek;
         this.baseSalary = (1000000 / 48) * this.activeHoursPerWeek;
     }
     public double getBaseSalary() { return this.baseSalary; }
 
-    public byte getActiveHoursPerWeek() { return this.activeHoursPerWeek;
+    public int getActiveHoursPerWeek() { return this.activeHoursPerWeek;
     }
 }
